Refactor getToken to utility function for other tests
